update qvae example
